100_zero_shot_with_assumption_java/09_src_function_analysis_ui/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
pyyaml==6.0.1
pydantic==2.5.2